LPS-71559 These classes are not exported, move to internal package
